reorganize functions definitions and declarations
